TechWeek_Hackathon2/requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
pandas
google-generativeai
markdown-it-py
Pillow
folium